Implement read-only root modeling context.
Previously the root modeling context was modifiable after TSEM
initialization had completed.  This commit implements the ability
to specify that the root modeling context be read only after TSEM
was initialized.

Previously, a static structure definition for the root modeling
context had been used.  Converting to a read-only modeling
context required that relevent members of the structure be
populated as part of the TSEM initialization process.

This work indicated that there was no need to have the root
modeling context or model be visible outside of the tsem.c file.
The variable declarations were moved from the namespace.c file to
the tsem.c file.