Enable generic modeling of unix socket events.
The following sockets events were not being modeled secondary to
the fact that they were causing system boots to hand:

unix_stream_connect

unix_may_send

These were events that were being called in atomic context and
the supposition was that extension into the TPM, that was a
blocking event, was causing a race that deadlocked the system.

With TPM extension now being done asynchronously the events were
tested and they no longer appear to be causing system hanges so
the generic modeling of the events was re-enabled.