Convert model state computation to run under a mutex.
Since the computation of the security state value was based on
the list of security state points the computation of the state
value had been converted to run under the spinlock that protected
list of security state values.  This was particularly problematic
given the amount of computation needed to compute the security
state value.

Following the model that has been used for taking the traversal
of the control plane list out from under a spinlock, the
computation of the security state point was converted from using
a spinlock to a mutex by only traversing the list of points that
were in effect when the request for computation of the value was
made.

There had previously been two lists of security state points
maintained, the primary list and a second sorted list, the latter
of which was used to compute the security state value for the
model.  This architecture was converted to a model where an array
of security event point descriptions was created based on the
active list of state points.  This array is sorted and used to
drive the generation of the security state value.

Since the array is only needed to compute the state value this
reduced, by half, the amount of memory needed to hold the
security state points for the model since the second sorted list
is not needed.  The memory for the array is vmalloced since the
generation of the state value is not a performance critical
operation.

A 'point_count' unsigned integer value was added to the
tsem_model structure in order to track the number of active
points in the model.  This value is used to drive the allocation
of the memory for the array of security state points.  This value
is protected by the spinlock that protects the list of security
state points and the size of the list is captured at the same
time as the end of the list for the computation of the state
value.