Restrict TMA tasks to only operate on their own namespace.
To further close the possibility of inserting an adversarial
process into the control sphere of an external trust orchestrator
the following member was added to struct tsem_task:

tma_for_ns

Creation of a namespace causes the new context id of the security
modeling namespace to be placed in this structure member.

When a trust orchestrator issues a call to set the trust status
of a task the security model context identifier of the target
process is compared to that of the task that is issueing the
call.  If they do not match EINVAL is returned as the result of
the control call.

To further strengthen the security guarantees the handling of
tsem_task_kill() was modified to change the permissions checking
that is done if the executing task has CAP_TRUST.  If the task is
running in the root modeling namespace permission is granted.  If
the task is a TMA for a modeling namespace the tma_for_ns member
of the TSEM task control structure must match the namespace
identifier for the target process.

This further limits the ability of an adversarial trust
orchestrator to attack a modeling namespace that is not under its
direct control.